SCRUM-291 refactor(posts): use api post types as source of truth
