+++ src/entities/posts/ui/PostModal/EditMode/EditModeImageSection/EditModeImageSection.tsx
@@ -1,9 +1,10 @@
 'use client'
+import type { PostViewModel } from '@/entities/posts/api'
+
 import React from 'react'
 
 import { Carousel } from '@/shared/composites'
 import { IMAGE_LOADING_STRATEGY, IMAGE_SIZES } from '@/shared/constant'
-import { PostViewModel } from '@/shared/types'
 import Image from 'next/image'
 
 import s from '../EditMode.module.scss'
