+++ src/shared/types/posts/models.ts
@@ -1,58 +1,5 @@
-import { UserImage } from '@/entities/users/api/api.types'
-
-import { LikeStatus } from '../base/enums'
 import { UserBase } from '../user/models'
 
-export interface PostImageViewModel {
-  url: string
-  width: number
-  height: number
-  fileSize: number
-  createdAt?: string
-  uploadId: string
-}
-
-export interface UploadedImageViewModel {
-  images: PostImageViewModel[]
-}
-
-export interface ChildMetadataDto {
-  uploadId: string
-}
-
-export interface CreatePostInputDto {
-  description?: string
-  childrenMetadata: ChildMetadataDto[]
-}
-
-export interface UpdatePostInputDto {
-  description?: string
-}
-
-export interface UpdateLikeStatusDto {
-  likeStatus: LikeStatus
-}
-
-export interface PostViewModel {
-  id: number
-  userName: string
-  description?: string
-  location?: string
-  images: UserImage[]
-  createdAt: string
-  updatedAt: string
-  ownerId: number
-  avatarOwner: string
-  owner: {
-    firstName: string
-    lastName: string
-  }
-  likesCount: number
-  isLiked: boolean
-  avatarWhoLikes: string[]
-}
-
-// Комментарии
 export interface BaseComment {
   id: number
   from: UserBase
@@ -73,16 +20,6 @@ export interface AnswersViewModel extends BaseComment {
 
 export interface CreateCommentDto {
   content: string
-}
-
-export interface PublicationsFollowersWithPaginationViewModel {
-  totalCount: number
-  pagesCount: number
-  page: number
-  pageSize: number
-  prevCursor: number
-  nextCursor: number
-  items: PostViewModel[]
 }
 
 export type PostVariant = 'public' | 'myPost' | 'userPost'
